fix(playlist): zero length playlist crash
this leads to another bug down the line... and the end user is in the
dark.

mitigation: exit as soon as the preconditions for the panic are met, in
this case, its checking if PLAYLIST is of length 0, and if that's the
case, then quit_with() a human readable error so the end user is not
left in the dark as to why it's not working.

bug was found by my tester, https://github.com/Vaneer420
love you <3