fix(sec): completely stomp out CVE-2021-29938
echotune was never affected by this, because minimp3-rs does not use the
offending functions. furthermore, echotune's `Cargo.toml` has the
`panic = "abort"` option in `Cargo.toml`, so the application will exit
on panic, not allowing the double free to ever occur.

this is a waste of time, but that crate (and some other deps it has) are
unmaintained, and i want my `cargo-audit audit` to look clean.

also 100th commit