fix: fix passing multiple audio files as args
imagine the arguments `./encore something.png something.mp3`
this won't work, because encore originally intreprets something.png as
a **non-audio** file, so it tries to parse it as a playlist, and dies
because &str/String expects to be utf8.
whereas... `./encore something.mp3 something.png` works right, because
encore treats it correctly (audio files instead of playlist)

this commit fixes that behaviour and makes it right, but due to how it
was written originally, it had to be refactored. its probably not the
best way to fix it, or rewrite it in this case, but it works.