fix(windows): probably fix playlists on windows
it just occured to me that we aren't using %userprofile% env var on
windows to find user profile dirs. nobody reported this as an issue
(since literally nobody puts their stuff in a playlist, and just runs
`./encore *`) but if it is, then its fixed now.

also adds a comment to clarify why we `continue` ing